fix(codegen): resolveLayout picks newest for unbounded impls + generated-tree gofmt (#17)
* fix(codegen): resolveLayout picks NEWEST layout when packetverMin is unbounded

PR #16 review acknowledged a latent codegen bug: encoders for actions with
an unbounded PacketverMin (min=0 in the semantic DB) emitted the OLDEST
layout variant instead of the newest. Concrete regression:
EncodeCzReqGuildEmblemImg2 produced a [10]byte encoder, but at goKore's
target packetver 20200401 rAthena expects 14 bytes (the 20190619+ variant
that adds a trailing 'unused' field — see packets_struct.hpp:5788-5803).

Root cause: resolveLayout treated PacketverMin=0 as 'pv=20030000', then
iterated ranges newest-first looking for the newest range with
'MinVer <= 20030000'. For structs whose earliest range starts after
20030000 (post-2019 packets in particular), no range matched, so the code
fell through to a forward-iteration fallback that returned the OLDEST
available range. Wrong direction.

Fix: when PacketverMin > 0, honor it (unchanged). When PacketverMin == 0
or no range covers the supplied min, fall through to the NEWEST available
range. Modern rAthena builds serve the newest struct variant on the wire,
so a single-variant encoder must emit that.

Adds 5 unit tests for resolveLayout: picks-newest-for-unbounded-impl (the
motivating case), honors-specific-packetverMin, packetverMin-below-all-ranges,
missing-struct, and unavailable-layouts.

Regenerated encoder output for cz_req_guild_emblem_img2 is now [14]byte
with the Unused field, matching rAthena at pv >= 20190619.

* chore(codegen): gofmt all generated .go files in writeFile

Templates for pv-branching encoders (see GenerateEncodeFile switch/case
path) emit correct-but-not-canonical whitespace: field-write statements
inside case bodies have extra tabs, and comment alignment isn't consistent
across sibling statements. Downstream this produced gofmt-dirty files that
either failed CI format checks or produced churn on unrelated PRs when
gofmt was run manually.

Add a format.Source() step to writeFile for any path ending in .go. If
gofmt fails (theoretically possible on very malformed templates), log a
warning and write unformatted — preserves historic fail-open behaviour.

This canonicalises the entire generated tree; most files touched by the
next commit are pure comment-alignment / const-indentation cosmetics.

* fix(semantics): bound cz_req_guild_emblem_img2 and split cz_req_takeoff_equip_all

Two mappings were unbounded despite having distinct packet-ID / layout
combinations at higher packetvers in rAthena:

- cz_req_guild_emblem_img2 (0x0B1E): bound to [20190227, ∞). The
  packets_struct.hpp guard is PACKETVER_MAIN_NUM >= 20190227, with a
  size-change at 20190619 that adds a trailing 'unused' field. Under
  MAIN-only semantics (per CONFIG_PACKETS_HPP block), the codegen
  correctly picks the newest layout for the impl.

- cz_req_takeoff_equip_all: split 0x0BAD [20210818, 20230905] + new
  0x0BF5 [20230906, ∞]. rAthena binds ZC_REQ_TAKEOFF_EQUIP_ALL to two
  DIFFERENT packet IDs across the pv=20230906 boundary AND changes the
  struct size (2 → 6 bytes with a Location field). Splitting into two
  impls lets the codegen emit a pv-branching encoder that pairs each
  layout with the correct packet ID on the wire.

Verified via g++ preprocess:
- pv=20200401: 0x0B1E binds to 14-byte variant (unused present).
- pv=20220101: 0x0BAD binds to 2-byte variant (no Location).
- pv=20240101: 0x0BF5 binds to 6-byte variant (Location present).

* chore(codegen): regenerate pkg/ outputs (gofmt canonicalization + two encoders)

Regenerated with:
- resolveLayout fix (encoders with unbounded impls now pick newest layout)
- writeFile gofmt step (canonicalises all generated Go files)
- cz_req_takeoff_equip_all split into 0x0BAD + 0x0BF5 impls

Net behavioural changes (small, described in the semantics/codegen commits):
- pkg/encode/cz_req_guild_emblem_img2.go: [10]byte → [14]byte with the
  Unused field. Correct for pv >= 20190619 (goKore's target pv=20200401).
- pkg/encode/cz_req_takeoff_equip_all.go: [6]byte → []byte with pv-branching.
  Correct for both 0x0BAD (pv 20210818..20230905, 2 bytes) and 0x0BF5
  (pv >= 20230906, 6 bytes with Location). goKore doesn't use this packet
  at pv=20200401 (doesn't exist that pv), so the API change is safe.
- pkg/encode/register.go: adapt to the new []byte return of
  EncodeCzReqTakeoffEquipAll (was passing b[:], now returns directly).

Everything else in this diff is pure gofmt cosmetics (comment alignment,
const-block indentation).

* docs: worklog 0091 + CHANGELOG [Unreleased] section

Documents the resolveLayout fix, generated-tree gofmt, and the two
mapping corrections. Cites the PR #16 review comment that acknowledged
this as a follow-up.

* fix(semantics): bind cz_req_guild_emblem_img2 at min=20190619 (actually 14 bytes now)

Reviewer of PR #17 caught that the previous mapping change (min=20190227)
did NOT produce a 14-byte encoder — resolveLayout's specific-pv path
picked the range [20190227, 20190619) which is the OLD 10-byte variant.
The worklog and commit message claimed [14]byte with Unused but the
actual output was still [10]byte.

rAthena's PACKET_CZ_REQ_GUILD_EMBLEM_IMG2 has two variants BOTH bound to
packet ID 0x0B1E (packets_struct.hpp:5788-5803):
  - pv 20190227..20190618 → 10 bytes (packetType + guild_id + emblem_id)
  - pv >= 20190619        → 14 bytes (adds trailing 'unused')

Since packet IDs are unique per action in the semantic DB, we cannot
express both variants as separate impls. We must pick one. Setting
min=20190619 corresponds to the CURRENT layout — what modern rAthena
builds (including goKore's target pv=20200401) emit on the wire.

The generated EncodeCzReqGuildEmblemImg2 is now [14]byte with the
Unused field, verified via a new wire-format test in
pkg/encode/regen_encoders_test.go. The narrow-window tradeoff (clients
at pv 20190227..20190618 would need a code change) is acceptable given
the low use of the packet and the ~4-month window.

Also adds encoder tests for cz_req_takeoff_equip_all pv-branching
behavior (per reviewer request):
  - pv=20210818: 2 bytes with 0x0BAD header
  - pv=20230906: 6 bytes with 0x0BF5 header + Location
  - pv=20200401: panics (packet doesn't exist)

Updated worklog to correct the false claims from the previous commit.
The resolveLayout newest-fallback is defense-in-depth for future
unbounded impls, not the mechanism that fixes CzReqGuildEmblemImg2.

Encoder tests added: TestEncodeCzReqGuildEmblemImg2_WireFormat,
TestEncodeCzReqTakeoffEquipAll_WireFormat_v20210818,
TestEncodeCzReqTakeoffEquipAll_WireFormat_v20230906,
TestEncodeCzReqTakeoffEquipAll_PanicsBelowRange.

* docs+bench: address APPROVE'd reviewer's non-blocking items

Three minor items flagged in the APPROVE review:

1. CHANGELOG line 44 said 'was [6]byte' — actual pre-PR type was [2]byte.
   Corrected.

2. CHANGELOG/worklog scope claimed cz_req_guild_emblem_img2 bound at
   [20190227, ∞) — actual binding is [20190619, ∞) (the pv where the
   14-byte layout becomes active). Corrected both docs; added a note
   explaining why (both variants share packet id 0x0B1E, so the
   semantic DB cannot express them as two impls — picking the current
   layout is the practical choice).

3. Added benchmarks for both changed encoders (codebase convention:
   25 encode benchmarks already exist). All three benchmarks hit
   0 allocs/op — even the []byte-returning EncodeCzReqTakeoffEquipAll,
   because Go's escape analysis proves the make([]byte, N) doesn't
   escape for these tiny-payload single-return cases. Rule 4 compliant.

+++ pkg/send/cz_req_random_combine_item.go
@@ -4,6 +4,6 @@ package send
 
 // CzReqRandomCombineItem is the request struct for the cz_req_random_combine_item action.
 type CzReqRandomCombineItem struct {
-	ItemId int16
-	Items []byte
+	ItemId int32
+	Items  []byte
 }
